--- src/main/java/com/reckue/account/configs/AuthServerConfig.java
@@ -1,117 +1,0 @@
-package com.reckue.account.configs;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.oauth2.config.annotation.configurers.ClientDetailsServiceConfigurer;
-import org.springframework.security.oauth2.config.annotation.web.configuration.AuthorizationServerConfigurerAdapter;
-import org.springframework.security.oauth2.config.annotation.web.configuration.EnableAuthorizationServer;
-import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerEndpointsConfigurer;
-import org.springframework.security.oauth2.config.annotation.web.configurers.AuthorizationServerSecurityConfigurer;
-import org.springframework.security.oauth2.provider.ClientDetailsService;
-import org.springframework.security.oauth2.provider.request.DefaultOAuth2RequestFactory;
-import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
-import org.springframework.security.oauth2.provider.token.TokenStore;
-import org.springframework.security.oauth2.provider.token.store.JwtAccessTokenConverter;
-import org.springframework.security.oauth2.provider.token.store.JwtTokenStore;
-
-import javax.annotation.PostConstruct;
-import java.util.Base64;
-
-/**
- * Class AuthServerConfig contains configurations as Authentication server.
- *
- * @author Kamila Meshcheryakova
- */
-@Configuration
-@RequiredArgsConstructor
-@EnableAuthorizationServer
-public class AuthServerConfig extends AuthorizationServerConfigurerAdapter {
-
-    @Value("${security.token.secret-key}")
-    private String secretKey;
-
-    private final PasswordEncoder passwordEncoder;
-    private final AuthenticationManager authenticationManager;
-    private final UserDetailsService userDetailsService;
-
-    @PostConstruct
-    protected void init() {
-        secretKey = Base64.getEncoder().encodeToString(secretKey.getBytes());
-    }
-
-    @Bean
-    public TokenStore tokenStore() {
-        return new JwtTokenStore(jwtAccessTokenConverter());
-    }
-
-    /**
-     * This method creates a Bean from our CustomTokenEnhancer as a access token converter
-     * which convert a token using secret key.
-     *
-     * @return converter
-     */
-    @Bean
-    public JwtAccessTokenConverter jwtAccessTokenConverter() {
-        var converter = new CustomTokenEnhancer();
-        converter.setSigningKey(secretKey);
-        return converter;
-    }
-
-    /**
-     * This method allows to set configurations for our client - account api in memory.
-     *
-     * @param clients information about clients of Authorization server
-     * @throws Exception all kind of checked exceptions
-     */
-    @Override
-    public void configure(ClientDetailsServiceConfigurer clients) throws Exception {
-        clients
-                .inMemory()
-                .withClient("client")
-                .secret(passwordEncoder.encode("secret"))
-                .authorizedGrantTypes("password", "authorization_token", "refresh_token")
-                .scopes("write");
-    }
-
-    @Bean
-    public DefaultTokenServices tokenServices() {
-        DefaultTokenServices defaultTokenServices = new DefaultTokenServices();
-        defaultTokenServices.setTokenStore(tokenStore());
-        defaultTokenServices.setSupportRefreshToken(true);
-        defaultTokenServices.setReuseRefreshToken(false);
-        return defaultTokenServices;
-    }
-
-//    @Autowired
-//    ClientDetailsService clientDetailsService;
-//
-//    @Bean
-//    DefaultOAuth2RequestFactory defaultOAuth2RequestFactory() {
-//        return //new DefaultOAuth2RequestFactory(clientDetailsService);
-//                new CustomOauth2RequestFactory(clientDetailsService);
-//    }
-
-    /**
-     * This method allows to set configurations for secured endpoints.
-     *
-     * @param endpoints that require a token
-     */
-    @Override
-    public void configure(AuthorizationServerEndpointsConfigurer endpoints) {
-        endpoints.authenticationManager(authenticationManager)
-                .userDetailsService(userDetailsService)
-                .tokenStore(tokenStore()).tokenEnhancer(jwtAccessTokenConverter())
-                .pathMapping("/oauth/token", "/auth/token");
-    }
-
-    @Override
-    public void configure(AuthorizationServerSecurityConfigurer security) throws Exception {
-        security.allowFormAuthenticationForClients();
-    }
-}
