+++ src/main/java/com/reckue/account/model/Status.java
@@ -1,0 +1,14 @@
+package com.reckue.account.model;
+
+/**
+ * Enum Status provides an enumeration of statuses for identifying user state.
+ *
+ * @author Kamila Meshcheryakova
+ */
+@SuppressWarnings("unused")
+public enum Status {
+
+    ACTIVE,
+    INACTIVE,
+    DELETED
+}
