--- src/main/java/com/reckue/account/models/Status.java
@@ -1,14 +1,0 @@
-package com.reckue.account.models;
-
-/**
- * Enum Status provides an enumeration of statuses for identifying user state.
- *
- * @author Kamila Meshcheryakova
- */
-@SuppressWarnings("unused")
-public enum Status {
-
-    ACTIVE,
-    INACTIVE,
-    DELETED
-}
