+++ src/main/java/com/reckue/account/transfer/RoleTransfer.java
@@ -1,0 +1,14 @@
+package com.reckue.account.transfer;
+
+import lombok.Data;
+
+/**
+ * Class RoleTransfer represents an outgoing DTO to receive a user role.
+ *
+ * @author Kamila Meshcheryakova
+ */
+@Data
+public class RoleTransfer {
+
+    private String name;
+}
