--- src/main/java/com/reckue/account/transfers/RoleTransfer.java
@@ -1,14 +1,0 @@
-package com.reckue.account.transfers;
-
-import lombok.Data;
-
-/**
- * Class RoleTransfer represents an outgoing DTO to receive a user role.
- *
- * @author Kamila Meshcheryakova
- */
-@Data
-public class RoleTransfer {
-
-    private String name;
-}
