+++ src/main/java/com/reckue/account/util/helper/TimestampHelper.java
@@ -1,0 +1,20 @@
+package com.reckue.account.util.helper;
+
+import java.sql.Timestamp;
+
+/**
+ * Class TimestampHelper represents a time generator.
+ *
+ * @author Kamila Meshcheryakova
+ */
+public class TimestampHelper {
+
+    /**
+     * This method is used to generate current time as long type.
+     *
+     * @return time as long type
+     */
+    public static long getCurrentTimestamp() {
+        return new Timestamp(System.currentTimeMillis()).getTime();
+    }
+}
