--- src/main/java/com/reckue/account/utils/helpers/TimestampHelper.java
@@ -1,20 +1,0 @@
-package com.reckue.account.utils.helpers;
-
-import java.sql.Timestamp;
-
-/**
- * Class TimestampHelper represents a time generator.
- *
- * @author Kamila Meshcheryakova
- */
-public class TimestampHelper {
-
-    /**
-     * This method is used to generate current time as long type.
-     *
-     * @return time as long type
-     */
-    public static long getCurrentTimestamp() {
-        return new Timestamp(System.currentTimeMillis()).getTime();
-    }
-}
